(PC-11890) feat(identityCheck) Add error handler for SetAddress
